# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EdgeEver API",
    "version": "0.1.0",
    "description": "REST API for EdgeEver notebooks, memos, resources, API tokens, and Agent integrations."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/ai/settings": {
      "get": {
        "summary": "List AI providers, their models, and the workspace default model",
        "responses": {
          "200": {
            "description": "AI settings with all API keys omitted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AiSettings" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/ai/providers": {
      "post": {
        "summary": "Add an AI provider service",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AiProviderCreateInput" } } }
        },
        "responses": {
          "201": { "description": "Provider created" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/ai/providers/{providerConfigId}": {
      "parameters": [{ "name": "providerConfigId", "in": "path", "required": true, "schema": { "type": "string" } }],
      "put": {
        "summary": "Update or enable/disable an AI provider service",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AiProviderUpdateInput" } } } },
        "responses": {
          "200": { "description": "Provider updated" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "summary": "Delete an AI provider and all of its models",
        "responses": {
          "200": { "description": "Provider deleted" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/ai/providers/{providerConfigId}/test": {
      "post": {
        "summary": "Test a saved AI provider with a model ID",
        "parameters": [{ "name": "providerConfigId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "modelId": { "type": "string", "minLength": 1, "maxLength": 200 } }, "required": ["modelId"] } } }
        },
        "responses": {
          "200": { "description": "The model endpoint responded successfully" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/ai/providers/{providerConfigId}/discover-models": {
      "post": {
        "summary": "Discover models exposed by a saved provider service",
        "parameters": [{ "name": "providerConfigId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Discovered models" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/ai/providers/{providerConfigId}/models": {
      "post": {
        "summary": "Add a model to an AI provider service",
        "parameters": [{ "name": "providerConfigId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AiModelCreateInput" } } } },
        "responses": {
          "201": { "description": "Model added" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/ai/providers/{providerConfigId}/models/{modelConfigId}": {
      "delete": {
        "summary": "Remove a model from an AI provider service",
        "parameters": [
          { "name": "providerConfigId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "modelConfigId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Model removed" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/ai/default-model": {
      "put": {
        "summary": "Choose or clear the workspace default AI model",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "modelConfigId": { "type": ["string", "null"] } }, "required": ["modelConfigId"] } } } },
        "responses": {
          "200": { "description": "Default model updated" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/ai/generate": {
      "post": {
        "summary": "Stream a note AI action using the current workspace model",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": { "type": "string", "enum": ["summarize", "extract-key-points", "extract-todos", "rewrite-proofread", "translate", "improve-writing", "fix-spelling-grammar", "make-shorter", "make-longer", "simplify-language", "change-tone", "continue-writing", "custom"] },
                  "promptId": { "type": "string", "minLength": 1, "maxLength": 200, "description": "Saved prompt to resolve inside the authenticated workspace" },
                  "locale": { "type": "string", "maxLength": 35 },
                  "title": { "type": "string", "maxLength": 160 },
                  "contentMarkdown": { "type": "string", "maxLength": 300000 },
                  "targetLanguage": { "type": "string", "enum": ["en", "zh-CN", "zh-TW", "ja", "ko", "es", "fr", "de", "pt"], "description": "Required when the action or selected prompt requests a target language" },
                  "tone": { "type": "string", "enum": ["professional", "friendly", "casual", "direct"] },
                  "instruction": { "type": "string", "minLength": 1, "maxLength": 2000, "description": "Freeform instruction; ignored when promptId is supplied" }
                },
                "required": ["action", "title", "contentMarkdown"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server-sent events containing start, text-delta, finish, or error events",
            "content": { "text/event-stream": { "schema": { "type": "string" } } }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "edgeever_session"
      }
    },
...
    "responses": {
...
      "Error": {
        "description": "API error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
...
      "AiModel": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "providerConfigId": { "type": "string" },
          "modelId": { "type": "string" },
          "displayName": { "type": "string" }
        },
        "required": ["id", "providerConfigId", "modelId", "displayName"]
      },
      "AiProvider": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "provider": { "type": "string", "enum": ["openai-compatible", "anthropic", "google"] },
          "displayName": { "type": "string" },
          "baseUrl": { "type": "string", "format": "uri" },
          "isEnabled": { "type": "boolean" },
          "hasApiKey": { "type": "boolean" },
          "models": { "type": "array", "items": { "$ref": "#/components/schemas/AiModel" } }
        },
        "required": ["id", "provider", "displayName", "baseUrl", "isEnabled", "hasApiKey", "models"]
      },
      "AiSettings": {
        "type": "object",
        "properties": {
          "providers": { "type": "array", "items": { "$ref": "#/components/schemas/AiProvider" } },
          "defaultModelId": { "type": ["string", "null"] },
          "encryptionConfigured": { "type": "boolean" },
          "readOnly": { "type": "boolean" }
        },
        "required": ["providers", "defaultModelId", "encryptionConfigured", "readOnly"]
      },
      "AiProviderCreateInput": {
        "type": "object",
        "properties": {
          "provider": { "type": "string", "enum": ["openai-compatible", "anthropic", "google"] },
          "displayName": { "type": "string", "minLength": 1, "maxLength": 80 },
          "baseUrl": { "type": "string", "format": "uri", "maxLength": 500 },
          "apiKey": { "type": "string", "minLength": 1, "maxLength": 4096 },
          "isEnabled": { "type": "boolean", "default": true },
          "initialModelId": { "type": "string", "minLength": 1, "maxLength": 200 }
        },
        "required": ["provider", "displayName", "baseUrl", "apiKey"]
      },
      "AiProviderUpdateInput": {
        "type": "object",
        "properties": {
          "provider": { "type": "string", "enum": ["openai-compatible", "anthropic", "google"] },
          "displayName": { "type": "string", "minLength": 1, "maxLength": 80 },
          "baseUrl": { "type": "string", "format": "uri", "maxLength": 500 },
          "apiKey": { "type": "string", "minLength": 1, "maxLength": 4096 },
          "isEnabled": { "type": "boolean", "default": true }
        },
        "required": ["provider", "displayName", "baseUrl"]
      },
      "AiModelCreateInput": {
        "type": "object",
        "properties": {
          "modelId": { "type": "string", "minLength": 1, "maxLength": 200 },
          "displayName": { "type": "string", "minLength": 1, "maxLength": 200 }
        },
        "required": ["modelId"]
      },
...
    }
  }
}